Update part 01 and part 02
Added :
1.	Introduction
This project is related to the delivery system that will be used by local delivery companies. We basically receive information from our customers, such as the weight and size of the shipment, and the destination of the delivery. Based on this information, the goal of this program is to find the truck with the best conditions and its route. Every truck has a fixed weight and volume. Each truck has its own truck with three different routes. Each truck can deliver through its designated route, or through space where no other truck's routes or buildings exist. Based on these conditions, the shortest possible distance for a truck is calculated. Compare the calculated distances to find the best route for the truck. If the distances are the same, we will find a truck that currently carries fewer shipments. Our test will produce our expected results by examining the conditions to be tested in this process, such as, 'Does this shortest distance pass through the space where a building exists?'
2.	Scope
a.	What will be tested.
i.	Whether the calculated route is crossing only the space that the truck can pass through
ii.	Whether the route reach a given destination
iii.	If the shipment is loaded, whether it does not exceed the maximum loading weight or volume
b.	What will not be tested.
i.	No test is conducted assuming a destination outside the 25 by 25 square grid